finally fixed the xX issue

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 21:47:58 by herbie            #+#    #+#             */
-/*   Updated: 2022/11/16 10:28:18 by codespace        ###   ########.fr       */
+/*   Updated: 2022/11/16 11:16:26 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,8 +31,8 @@ int	ft_intlen(int n);
 char	*ft_strchr(const char *s, int c);
 int	ft_putnbr_fd(int n, int fd);
 int	ft_putunbr_fd(unsigned int n, int fd);
-int	ft_puthex_fd(unsigned long n, int fd);
-int	ft_puthex_up_fd(long long n, int fd);
+int	ft_puthex_fd(unsigned int n, int fd);
+int	ft_puthex_up_fd(unsigned int n, int fd);
 int ft_putptr_fd(unsigned long n, int fd);
 int	ft_putnbr_base_fd(unsigned long n, char *base, int fd);
 int	ft_putstr_fd(char *s, int fd);
