+++ ft_printf/ft_puthex_fd.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_puthex_fd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: herbie <herbie@student.42.fr>              +#+  +:+       +#+        */
+/*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/14 16:50:33 by herbie            #+#    #+#             */
-/*   Updated: 2022/11/16 09:59:02 by herbie           ###   ########.fr       */
+/*   Updated: 2022/11/16 11:16:53 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_puthex_fd(unsigned long n, int fd)
+int	ft_puthex_fd(unsigned int n, int fd)
 {
-	return (ft_putnbr_base_fd(n, "0123456789abcdef", fd));
+	return (ft_putnbr_base_fd((unsigned long)n, "0123456789abcdef", fd));
 }
